I wrote a methode to check diseases with agents new immuneSystem.

+++ LastSugarFaz7/src/Interfaces/ISpace_Diseases.java
@@ -4,5 +4,5 @@ import java.util.ArrayList;
 
 public interface ISpace_Diseases extends ISpaceProvider{
 
-    public ArrayList<Short> getDiseases();
+    public ArrayList<Integer> getDiseases();
 }
